% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendFFCols.R
\name{appendFFCols}
\alias{appendFFCols}
\title{This function is a slightly adjusted copy of the fr_append_cols
function in flowCore. It mainly allows for the presence of additional
pData variables}
\usage{
appendFFCols(focusFrame, newCols)
}
\arguments{
\item{focusFrame}{The flowFrame that will get appended.}

\item{newCols}{The new columns that should be added.}
}
\value{
The focusFrame, with the newCols appended.
}
\description{
This function is a slightly adjusted copy of the fr_append_cols
function in flowCore. It mainly allows for the presence of additional
pData variables
}
