% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrMatCreate.R
\name{corrMatCreate}
\alias{corrMatCreate}
\title{Generate a correction matrix for cytometry data analysis}
\usage{
corrMatCreate(specMat)
}
\arguments{
\item{specMat}{The spectral matrix used to unmix the dataset of interest.}
}
\value{
A symmetrical matrix of zeros with the right row- and column names.
}
\description{
This function aids the correctUnmix function, to create a symmetrical
correction matrix that should be used together with a flowframe to correct
the errors of unmixing.
}
\seealso{
\code{\link{correctUnmix}}
}
