% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOut.R
\name{filterOut}
\alias{filterOut}
\title{Select events based on filter}
\usage{
filterOut(flowObj, filterName, withinFilter = TRUE)
}
\arguments{
\item{flowObj}{The fcs object to be subsetted. Both flowFrames and flowSets
are accepted.}

\item{filterName}{The name of the filter variable in the flow object to be
used for the subsetting.}

\item{withinFilter}{Should the events within or outsitde of the filter be
selected?}
}
\value{
The flowObj, now smaller, as it only contains the data within,
or outside of the filterName filter.
}
\description{
Select events based on filter
}
