% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullPanel.R
\docType{data}
\name{fullPanel}
\alias{fullPanel}
\title{A fully stained spectral cytometry sample}
\format{An object of class \code{"flowFrame"}}
\usage{
data(fullPanel)
}
\description{
This is a flowFrame with a PBMC sample stained with 12 fluorochromes.
Data acquired on a 44 detector, 3 laser Cytek Aurora® instrument by
J Theorell.
Date: 2018-10-25.
}
\keyword{datasets}
