% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendFFCols.R, R/newExprs.R
\name{newExprs<-}
\alias{newExprs<-}
\title{This function lets us exchange a flowframes exprs portion to an unrelated one
It is solely meant to be used within appendFFCols.}
\usage{
newExprs(x) <- value

newExprs(x) <- value
}
\arguments{
\item{x}{A flowFrame}

\item{value}{A matrix suitable to be an exprs object.}
}
\value{
A new flowFrame with "value" as the exprs portion.
}
\description{
This function lets us exchange a flowframes exprs portion to an unrelated one
It is solely meant to be used within appendFFCols.

This function lets us exchange a flowframes exprs portion to an unrelated one
}
\keyword{internal}
