% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDownSample.R
\name{plotDownSample}
\alias{plotDownSample}
\title{This function is a co-function to the oneVsAllPlot function, to reduce
computational time.}
\usage{
plotDownSample(flowObj, nRows = 10000)
}
\arguments{
\item{flowObj}{The object to downsample}

\item{nRows}{The number of rows in the downsampled sample.}
}
\value{
The downsampled flowObj.
}
\description{
This function is a co-function to the oneVsAllPlot function, to reduce
computational time.
}
\keyword{internal}
