% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specMat.R
\docType{data}
\name{specMat}
\alias{specMat}
\title{Spectral unmixing matrix}
\format{An object of class \code{"matrix"};}
\usage{
data(specMat)
}
\description{
This matrix is generated using the specMatCalc function and the unmixCtrls
example file.
}
\keyword{datasets}
