% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSpecs-package.R
\docType{package}
\name{flowSpecs-package}
\alias{flowSpecs-package}
\alias{flowSpecs}
\title{Tools for processing high-dimensional cytometry files}
\description{
This package is intended to fill the role of conventional
cytometry pre-processing software, for spectral decomposition,
transformation, visualization and cleanup, and to aid further downstream
analyses, such as with DepecheR, by enabling transformation of flowFrames
and flowSets to dataframes. Functions for flowCore-compliant automatic
1D-gating/filtering are in the pipe line.
}
\seealso{
\code{\link[flowCore:flowCore-package]{flowCore}}
}
\author{
Maintainer: Jakob Theorell <jakob.theorell@ndcn.ox.ac.uk>
}
\keyword{package}
